feedback partially addressed
